Create String With Numbers At Its End
Given string s that is appended with a number at last. The task is to check whether the length of string excluding that number is equal to that number.
Example 1:

Input:  s = "geeks5"
Output: 1
Explanation: Length of geeks is 5
and the last number is also 5.
Example 2:

Input:  s = "geek5"
Output: 0
Explanation: Length of geek is 4
and the last number is 5.
User Task:
Your task is to complete the function isSame() which takes a single string as input and returns 1 in case the answer is true, otherwise 0. You do not need to take any input or print anything.

Expected Time Complexity: O(|s|)
Expected Auxiliary Space: O(1)